+++ Sources/Navigator/EPUB/Scripts/webpack.config.js
@@ -6,10 +6,25 @@ module.exports = {
   entry: {
     reflowable: "./src/index-reflowable.js",
     fixed: "./src/index-fixed.js",
-    "fixed-wrapper": "./src/index-fixed-wrapper.js",
+    "fixed-wrapper-one": "./src/index-fixed-wrapper-one.js",
+    "fixed-wrapper-two": "./src/index-fixed-wrapper-two.js",
   },
   output: {
     filename: "readium-[name].js",
     path: path.resolve(__dirname, "../Assets/Static/scripts"),
+  },
+  module: {
+    rules: [
+      {
+        test: /\.m?js$/,
+        exclude: /node_modules/,
+        use: {
+          loader: "babel-loader",
+          options: {
+            presets: ["@babel/preset-env"],
+          },
+        },
+      },
+    ],
   },
 };